if statement checking, better parsing, sections fix & signals
If statements are now checked to see if they can evaluate properly. Empty lines are now allowed in spell code. Multiple sections can now be used at the same time (their was a bug where you couldn't). The setup for sending a signal when a component is called has been done but isn't ready and will most likely change.